Small refactoring and code style changes
* Use size_t/ssize_t data types where appropiate

* renamed swoole_hex2dec test to hex2dec

* use Event instead of swEvent

* Use SW_LOOP_N macro for loop

* use Server instead of swServer

* use ListenPort instead of swListenPort

* renamed Server::set_ipc_max_size to Server::init_ipc_max_size and renamed swIPC_type to swIPCMode

+++ include/swoole.h
@@ -665,7 +665,7 @@ struct Global {
     network::Socket *aio_default_socket;
     //-----------------------[Hook]--------------------------
     void *hooks[SW_MAX_HOOK_TYPE];
-    std::function<bool(Reactor *reactor, int &event_num)> user_exit_condition;
+    std::function<bool(Reactor *reactor, size_t &event_num)> user_exit_condition;
 
     // bug report message
     std::string bug_report_message = "";
